BAN-KVQA/ban_requirements.txt
backcall==0.2.0
boto3==1.23.10
botocore==1.26.10
certifi==2023.7.22
charset-normalizer==2.0.12
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
h5py==2.9.0
idna==3.4
imageio==2.15.0
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
jmespath==0.10.0
joblib==1.1.1
JPype1==0.7.0
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.3.1
konlpy==0.5.1
matplotlib==3.3.4
mecab-python===0.996-ko-0.9.2
nest-asyncio==1.5.8
numpy==1.16.3
opencv-python==4.5.4.60
pandas==1.1.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
progressbar2==3.55.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
Pygments==2.14.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-utils==3.5.2
pytorch-pretrained-bert==0.6.2
pytz==2023.3.post1
pyzmq==25.1.1
regex==2023.8.8
requests==2.27.1
s3transfer==0.5.2
scikit-learn==0.24.2
scipy==1.5.4
sentencepiece==0.1.83
six==1.16.0
threadpoolctl==3.1.0
torch==1.1.0
torchvision==0.3.0
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.18
wcwidth==0.2.10
zipp==3.6.0
